rename files and directories in components
