+++ src/components/case_manager/AllParas.tsx
