+++ src/pages/cmDashboard.tsx
@@ -1,6 +1,6 @@
 import React from "react";
-import AllParasPage from "../components/CaseManagerUI/paras";
-import AllStudentsPage from "../components/CaseManagerUI/students";
+import AllParasPage from "../components/case_manager/AllParas";
+import AllStudentsPage from "../components/case_manager/AllStudents";
 import styles from "../styles/Home.module.css";
 import Head from "next/head";
 import Image from "next/image";
